requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1
openai>=1.40.0
